vault, vault-bin: add CVE-2024-2660 to the knownVulnerabilities
Ideally we would upgrade but in this specific instance the fixed
versions (and the patch for the issue) have been relicensed under
the Business Source License which is unfree.

Tagging the packages with `knownVulnerabilities` seems to be our least
worse solution. If someone care enough I'm guessing the unfree versions
could be backported under a different names in 23.11.