octoprint: apply patch for CVE-2024-28237
The 1.10.0 upgrade from #306861 seems to be too big to
be safely backported.

I also did not backport CVE-2024-23637: it is split across
a series of patches and impacts quite sensitive parts of OctoPrint.
I am not feeling confident enough to backport it and exploiting the issue
requires an admin level access.